Basic Implementation of Netlify CMS
Not finished and not working, yet. They recommend the "Git Gateway" backend, which requires us to use their Netlify Identity product, but I'd like to avoid that, as it means we'll have to manage users separately.

They offer a github-oauth-api gateway which we can use to get things started, and later switch to a self-hosted proxy, if we prefer, without much hassle.

Also, I only added the config for the news collection, so far.