NetlifyCMS: Add /funktionen
This is still very WIP. The sections show but not all fields or everything is working.

Still needs to allow editing the full `sections` and then have a better preview.

I also noticed the sections/waterfall template hasn't been migrated to nunjucks and was being ignored, so I'll also fix that.

Fixes #41